refactor: move the strategy helpers into a dedicated module
