+++ src/core/strategies/index.js
@@ -16,11 +16,8 @@
  * Copyright (c) 2023 Open Assessment Technologies SA ;
  */
 
-export { default as engineFactory } from './engine.js';
-export { default as mathsEvaluatorFactory } from './mathsEvaluator.js';
-export { default as tokenizerFactory } from './tokenizer.js';
-export { default as expressionHelper, defaultDecimalDigits } from './expression.js';
-export { default as tokensHelper } from './tokens.js';
 /* c8 ignore start */
-export * from './terms.js';
+export * from './helpers.js';
+export * from './token.js';
+export * from './value.js';
 /* c8 ignore end */
